requirements.txt
# Simulation de Sable 3D - Dépendances
numpy>=1.24.0
PyQt6>=6.5.0
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
pyrr>=0.10.3
